ci: add sla e2e test to ensure functional form dwctl refactor: move test_utils into test module and add all e2e tests there
