+++ dwctl/Cargo.toml
@@ -16,11 +16,10 @@ path = "src/main.rs"
 [features]
 default = ["embedded-db"]
 embedded-db = ["dep:postgresql_embedded"]
-test-utils = ["dep:axum-test"]
 
 [dependencies]
 axum = { version = "0.8", features = ["multipart"] }
-fusillade = { version = "0.12.0" }
+fusillade = { version = "0.12.1" }
 tokio = { version = "1.0", features = ["full"] }
 tokio-stream = { version = "0.1", features = ["sync"] }
 tokio-util = "0.7"
@@ -98,8 +97,6 @@ postgresql_embedded = { version = "0.20", optional = true, features = [
 rust-embed = { version = "8.5", features = ["debug-embed"] }
 mime_guess = "2.0"
 futures = "0.3.31"
-# axum-test is optional and only used for tests and benchmarks
-axum-test = { version = "17.3", optional = true }
 dashmap = { version = "6.1.0", features = ["serde"] }
 moka = { version = "0.12", features = ["future"] }
 # Stripe payment processing
@@ -112,7 +109,7 @@ async-stripe = { version = "0.41.0", default-features = false, features = [
 ] }
 
 [dev-dependencies]
-axum-test = { version = "17.3" }
+axum-test = { version = "18.4.1" }
 sqlx = { version = "0.8", features = [
   "postgres",
   "runtime-tokio-rustls",
