+++ dwctl/src/api/handlers/api_keys.rs
@@ -364,7 +364,7 @@ mod tests {
     use crate::api::models::api_keys::{ApiKeyInfoResponse, ApiKeyResponse};
     use crate::api::models::pagination::PaginatedResponse;
     use crate::api::models::users::Role;
-    use crate::test_utils::*;
+    use crate::test::utils::*;
     use serde_json::json;
     use sqlx::PgPool;
